refactoring for Post and Timeline components, subdividing each of them into smaller components.
