+++ src/components/Container.js
@@ -3,7 +3,7 @@ import { containerPadding} from "../utils/MarginAndPaddingUtils"
 
 const Container = styled.main`
     width: 100%;
-    min-height: 100vh;
+    min-height: 101vh;
     padding: 72px ${containerPadding(611,40,301)} 40px;
     background-color: #333333;
     display: flex;
