+++ src/contexts/PostContext.js
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+const PostContext = createContext();
+
+export default PostContext;
