+++ src/utils/MarginAndPaddingUtils.js
@@ -1,5 +1,5 @@
 const containerPadding = (postWidth, widthBetweenPostAndTrending, trendingWidth) => {
-    const padding = `calc( ( 100vw - ${ postWidth }px - ${ widthBetweenPostAndTrending }px - ${ trendingWidth }px ) / 2)`;
+    const padding = `calc( ( 100% - ${ postWidth }px - ${ widthBetweenPostAndTrending }px - ${ trendingWidth }px ) / 2)`;
     return padding;
 }
